% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segMINT_db.R
\name{get_results}
\alias{get_results}
\title{Get simulation results from RDS files}
\usage{
get_results(dir, N = "all")
}
\arguments{
\item{dir}{Character string. Path to directory containing simulation_results_*.rds files}

\item{N}{Either "all" to process all files, or an integer to process files 1:N}
}
\value{
A data frame containing processed simulation results with flattened parameters
}
\description{
Reads and processes simulation result files from a specified directory
}
\examples{
\dontrun{
results <- get_results("/path/to/data", "all")
results <- get_results("/path/to/data", 100)
}
}
